// src/components/Interfaces.ts
export interface PokemonI {
  name: string
  id: string
}
export interface RegionI {
  name: string
  id: string
}

export interface IconsI {
  type: string,
  icon: string,
  color: string,
  id: Number
}

export interface InformationI {
  name: string
  id: string
}

export interface TypesI {
  type: string
  id: string
  icon: string
  color: string
}